requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
openai>=1.0.0
python-dotenv>=1.0.0
gtts>=2.3.0
ollama>=0.1.0
requests>=2.31.0
